% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{apply_effects}
\alias{apply_effects}
\title{apply_effects}
\usage{
apply_effects(inputs, image)
}
\arguments{
\item{inputs}{A list of valid `magick` effects}

\item{image}{A object returned by `magick::image_read`}
}
\value{
A pointer to a 'magick-image' object
}
\description{
Takes a shiny input and applies `magick` filters based on those inputs
}
\keyword{internal}
