% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_row_module.R
\name{deletePointUI}
\alias{deletePointUI}
\title{deletePointUI}
\usage{
deletePointUI(id, label = "Delete Last Point")
}
\description{
UI for the module that removes the last row in the dataframe of clicks.
}
\keyword{internal}
