% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DMMVI.R
\name{DMMVI}
\alias{DMMVI}
\title{DMMVI}
\usage{
DMMVI(X, type, A, H, eta, nu, alpha, lam, mu0, sig0, pr0, alpha_sigma,
  beta_sigma, alpha_p, beta_p, binit = NULL)
}
\arguments{
\item{X}{n by p matrix where n is the sample size and p is the feature size}

\item{type}{p-dimensional vector of feature types. Currently supports continuous(=0) and binary(=1)}

\item{A}{the adjacency matrix of the graph representing the graphical structure of X}

\item{H}{the number of clusters to be fit}

\item{eta}{sparsity parameter for gamma's}

\item{nu}{smoothness parameter for gamma's}

\item{alpha}{concentration parameter for dirichlet process}

\item{lam}{NEED DESCRIPTION}

\item{mu0}{mean of the non-selected continuous features}

\item{sig0}{variance of the non-selected continuous features}

\item{pr0}{"on" probability of the non-selected binary features}

\item{alpha_sigma}{shape parameter of the prior of residual variance(sigma^2)}

\item{beta_sigma}{rate parameter of the prior of residual variance(sigma^2)}

\item{alpha_p}{first shape parameter of the prior of the "on" probabilities(p_hj) of binary features}

\item{beta_p}{second shape parameter of the prior of the "on" probabilities(p_hj) of binary features}

\item{binit}{n by H initial matrix of B, exp(B_ih) is proportional to Pr(z_i=h). If NULL (default), random numbers are filled in.}
}
\description{
Variational Inference for Dirichlet Mixture Model
(ver. 20191126)
}
\examples{
# ADD EXAMPLE
}
\author{
Changgee Chang
}
