% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_sample_seurat}
\alias{load_sample_seurat}
\title{Load sample_seurat example data}
\usage{
load_sample_seurat()
}
\value{
A Seurat object containing example single-cell data
}
\description{
Loads the sample_seurat example dataset for testing the scToNeb function.
This dataset contains a subset (1000 genes and 1000 cells) of the eight-pancreas scRNA-seq datasets.
}
\details{
Load sample_seurat data

This function requires the Seurat package to be installed.
If Seurat is not available, the function will throw an error with installation instructions.
}
\examples{
\dontrun{
library(nebula)
sample_seurat <- load_sample_seurat()
re <- scToNeb(obj = sample_seurat, assay = "RNA", id = "replicate",
              pred = c("celltype", "tech"))
}

}
\seealso{
\code{\link{scToNeb}}
}
\keyword{data}
