% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanCancer-package.R
\docType{package}
\name{Nebula-package}
\alias{Nebula}
\alias{Nebula-package}
\title{Nebula: Network-based latent-Dirichlet subtype analysis (Nebula)}
\description{
Functions implementing the Network-based latent-Dirichlet subtype analysis (Nebula) algorithm for variational inference using Dirichlet mixture model with incorporation of graphical knowledge. Practically, this can be used to incorporate biological networks/pathways to inform clustering solutions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/changgee/Nebula}
  \item Report bugs at \url{https://github.com/changgee/PanCancer/issues}
}

}
\author{
\strong{Maintainer}: Changgee Chang \email{changgee@pennmedicine.upenn.edu}

Authors:
\itemize{
  \item Yize Zhao \email{yize.zhao@yale.edu}
  \item Margaret Hannum \email{hannumm@mskcc.org}
  \item Ronglai Shen \email{shenr@mskcc.org}
}

}
\keyword{internal}
