% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NebulaCore.R
\name{NebulaCore}
\alias{NebulaCore}
\title{NebulaCore}
\usage{
NebulaCore(
  X,
  type,
  E,
  H,
  eta,
  nu,
  alpha,
  lam,
  alpha_sigma,
  beta_sigma,
  alpha_p,
  beta_p,
  mu0,
  sig0,
  pr0,
  binit = NULL
)
}
\arguments{
\item{X}{n by p matrix where n is the sample size and p is the feature size}

\item{type}{p-dimensional vector of feature types. Currently supports continuous(=0) and binary(=1)}

\item{E}{e by 2 matrix of the graph edges representing the graphical structure of X}

\item{H}{the number of clusters to be fit}

\item{eta}{length p vector of sparsity parameters}

\item{nu}{smoothness parameter for gamma's}

\item{alpha}{concentration parameter for dirichlet process}

\item{lam}{shrinkage parameter for means of selected continuous features}

\item{alpha_sigma}{shape parameter of the prior of residual variance(sigma^2)}

\item{beta_sigma}{rate parameter of the prior of residual variance(sigma^2)}

\item{alpha_p}{first shape parameter of the prior of the 'active' probabilities(p_hj) of binary features}

\item{beta_p}{second shape parameter of the prior of the 'active' probabilities(p_hj) of binary features}

\item{mu0}{mean of the non-selected continuous features}

\item{sig0}{variance of the non-selected continuous features}

\item{pr0}{'active' probability of the non-selected binary features}

\item{binit}{n by H initial matrix of B, exp(B_ih) is proportional to Pr(z_i=h). If NULL (default), random numbers are filled in.}
}
\description{
Network based latent dirichlet subtype analysis
(ver. 20190920)
}
\examples{
# ADD EXAMPLE
}
\author{
Changgee Chang
}
