% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_row.R
\name{remove_last_row}
\alias{remove_last_row}
\title{remove_last_row}
\usage{
remove_last_row(click_data_reactive)
}
\arguments{
\item{click_data_reactive}{}
}
\value{
A new reactive dataframe, with the last row removed
}
\description{
Removes the last row from a reactive dataframe
}
\keyword{internal}
