% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_module.R
\name{ResetButtonModule}
\alias{ResetButtonModule}
\alias{resetButtonUI}
\alias{resetButton}
\alias{confirmDeleteUI}
\alias{confirmDelete}
\title{module definitions for the reset button module}
\usage{
resetButtonUI(id, label = "Reset All")

resetButton(input, output, session)

confirmDeleteUI(id, label = "Delete all Points")

confirmDelete(input, output, session)
}
\description{
This module starts a modal that confirms whether the user wants to reset nebula back to its original state when loaded. This is useful when you have finished analysing a otolith image, and want to open a new one.

Confirmation dialog for delete
}
\section{resetButtonServer}{

creates a secondary ui asking for confirmatio
}

\keyword{internal}
