% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_module.R
\name{resetButtonUI}
\alias{resetButtonUI}
\title{module definitions for the reset button module}
\usage{
resetButtonUI(id, label = "Reset All")
}
\description{
This module starts a modal that confirms whether the user wants to reset nebula back to its original state when loaded. This is useful when you have finished analysing a otolith image, and want to open a new one.
}
\keyword{internal}
