% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_seurat}
\alias{sample_seurat}
\title{An example Seurat object for testing scToNeb}
\format{
A Seurat object
}
\source{
\url{https://github.com/satijalab/seurat-data}
}
\description{
A Seurat object containing a subset (1000 genes and 1000 cells) of the eight-pancreas scRNA-seq datasets for testing the scToNeb function.
}
\details{
An example Seurat object for testing scToNeb
}
\keyword{datasets}
