% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{show_click}
\alias{show_click}
\title{show_click}
\usage{
show_click(click_data, x_val, y_val)
}
\arguments{
\item{click_data}{Reactive dataframe with x and y values}

\item{x_val}{Plot input for single click interaction x values}

\item{y_val}{Plot input for single click interaction y values}
}
\description{
Process the plot single click and stores in dataframe. Used within an observeEvent
}
\keyword{internal}
