% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_interactions.R
\name{show_transect}
\alias{show_transect}
\title{show_transect}
\usage{
show_transect(transect_dat, x_val, y_val, input_)
}
\arguments{
\item{transect_dat}{Reactive dataframe holding x and y values}

\item{x_val}{Plot input for double click interaction x values}

\item{y_val}{Plot input for double click interaction y values}

\item{input_}{Placeholder parameter to pass the `input$hover` from enclosing env}
}
\description{
Processes the plots double click and hover inputs to create a transect line. Used within an observeEvent
}
\keyword{internal}
