% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Nebula-package.R
\docType{package}
\name{nebula-package}
\alias{nebula-package}
\alias{_PACKAGE}
\alias{Nebula-package}
\title{nebula: Network-Based Latent-Dirichlet Subtype Analysis}
\description{
Functions implementing the Network-based latent-Dirichlet subtype analysis (nebula) algorithm for variational inference using Dirichlet mixture model with incorporation of graphical knowledge. Practically, this can be used to incorporate biological networks/pathways to inform clustering solutions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nebula-group/nebula}
  \item Report bugs at \url{https://github.com/nebula-group/nebula/issues}
}

}
\author{
\strong{Maintainer}: Margaret Hannum \email{margie.hannum@gmail.com}

Authors:
\itemize{
  \item Changgee Chang \email{changgee@pennmedicine.upenn.edu}
  \item Yize Zhao \email{yize.zhao@yale.edu}
  \item Ronglai Shen \email{shenr@mskcc.org}
}

}
\keyword{internal}
