% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nebula-package.R, R/run_app.R
\docType{package}
\name{nebula}
\alias{nebula}
\alias{_PACKAGE}
\alias{nebula-package}
\title{nebula: What the Package Does (Title Case)}
\usage{
nebula()
}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.

Starts the Otolith analysis app in your web browser
}
\examples{
\dontrun{
nebula()
}
}
\keyword{internal}
