% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PanCancer-package.R
\docType{package}
\name{PanCancer-package}
\alias{PanCancer}
\alias{PanCancer-package}
\title{PanCancer: Variational Inference for Dirichlet Mixture Model with Incorporation of Graphical Knowledge}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/changgee/PanCancer}
  \item Report bugs at \url{https://github.com/changgee/PanCancer/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (YOUR-ORCID-ID)

}
\keyword{internal}
