% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_data}
\alias{sample_data}
\title{An example data set for testing nebula}
\format{
A list of four objects:
\describe{
  \item{count}{A raw count matrix}
  \item{sid}{A vector of subject IDs}
  \item{pred}{A data frame of three predictors}
  \item{offset}{A vector of scaling factors}
}
}
\usage{
sample_data
}
\description{
A dataset containing a count matrix, subject IDs, a data frame of predictors and scaling factors.
}
\keyword{datasets}
