\name{sample_seurat}
\alias{sample_seurat}
\docType{data}
\title{
An example data set for testing scToNeb
}
\description{
A Seurat object containing a subset (1000 genes and 1000 cells) of the eight-pancreas scRNA-seq datasets.
}
\usage{data("sample_seurat")}
\format{
  A Seurat object
}

\source{
https://github.com/satijalab/seurat-data
}

\examples{
data(sample_seurat)
}
\keyword{datasets}
